requirements.txt
gum-ai
mcp[cli]